+++ src/main/kotlin/racingcar/Car.kt
@@ -1,21 +1,25 @@
 package racingcar
 
-import racingcar.util.ErrorMessage
-
-class Car private constructor(
+class Car(
     val name: String,
     initialDistance: Int = 0,
 ) {
     var distance: Int = initialDistance
         private set
 
-    fun move(number: Int) {
-        if (isMovable(number)) {
+    fun move() {
+        val random: Int = (0..9).random()
+        if (isMovable(random)) {
             distance += 1
         }
+        displayDistance()
     }
 
-    private fun isMovable(random: Int): Boolean = random >= 4
+    private fun displayDistance() {
+        var prettyDistance = ""
+        repeat(distance) { prettyDistance += "-" }
+        println("$name : $prettyDistance")
+    }
 
     override fun hashCode(): Int {
         var result = name.hashCode()
@@ -36,19 +40,9 @@ class Car private constructor(
     }
 
     companion object {
-        const val CAR_NAME_SEPARATOR: String = ","
-        const val MAX_CAR_NAME_LENGTH: Int = 5
-
-        /**
-         * create List<Car> from string value.
-         * @throws IllegalArgumentException when any car name doesn't follow car name rule
-         * **/
-        fun createCars(value: String): List<Car> {
-            val carNames: List<String> = value.split(CAR_NAME_SEPARATOR).map { it.trim() }
-            if (carNames.any { carName: String -> carName.length > MAX_CAR_NAME_LENGTH }) {
-                throw IllegalArgumentException(ErrorMessage.OVER_MAX_CAR_NAME_LENGTH)
-            }
-            return carNames.map { carName: String -> Car(carName) }
+        fun isMovable(random: Int): Boolean {
+            // 전진하는 조건은 0에서 9 사이에서 무작위 값을 구한 후 무작위 값이 4 이상일 경우이다.
+            return random >= 4
         }
     }
 }
