+++ src/test/kotlin/ApplicationTest.kt
@@ -1,5 +1,7 @@
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
+import racingcar.Car
+import racingcar.Racecourse
 
 class ApplicationTest {
     @Test
